# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HKTVMall Grocery Price Tracker API
 * OpenAPI spec version: 0.3.0
 */

export interface Alert {
  id: number;
  productId: number;
  productName: string;
  targetPrice: number;
  currentPrice: number;
  isTriggered: boolean;
  createdAt: Date;
}