# lib/api-zod/src/generated/types/createAlertRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HKTVMall Grocery Price Tracker API
 * OpenAPI spec version: 0.3.0
 */

export interface CreateAlertRequest {
  productId: number;
  targetPrice: number;
}