# lib/api-zod/src/generated/types/productDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HKTVMall Grocery Price Tracker API
 * OpenAPI spec version: 0.3.0
 */
import type { PriceRecord } from "./priceRecord";
import type { Product } from "./product";

export type ProductDetail = Product & {
  priceHistory: PriceRecord[];
};