# lib/api-zod/src/generated/types/refreshResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HKTVMall Grocery Price Tracker API
 * OpenAPI spec version: 0.3.0
 */

export interface RefreshResult {
  refreshed: number;
  failed: number;
  priceDrops: number;
  errors: string[];
}